refactor(inventory): split ItemForm into a use-item-form hook
